Post-push fix for BUG#20230459
The CTRL+C was original enabled but would only work if there
were non-daemon threads running. Otherwise, the main thread
would get the notification but the program would not abort
waiting for the non-daemon threads to finish.

In BUG#20230459, we have disabled CTRL+C but we got requests
to re-enabled it. In this follow up patch, we have done so.

Users willing to abort the MySQL Fabric process will have to
press CTRL+C three times now after which the process will be
aborted and exit with the error code 130 (Script terminated
by Control-C).

Note that the MySQL Fabric process will abruptly exit without
waiting wait for non-daemon threads and each time the CTRL+C
is pressed the following message will be printed out:

  "Request {0} to abort process. Please, send {1} additional "
  "request(s).".format(attempts, 3 - attempts))